Separate story list and story test cases with shared helper methods
